提交现勘5.0版本定时获取案件号的功能，修改优化部分代码，以及提交一部分现勘相关的单元测试代码

+++ hall-webservice/src/main/scala/nirvana/hall/webservice/internal/survey/gafis62/FPT50HandprintServiceCron.scala
@@ -113,7 +113,7 @@ class FPT50HandprintServiceCron(hallWebserviceConfig: HallWebserviceConfig,
     //获取待发送现场指掌纹数量
     val latentCount = fPT50HandprintServiceClient.getLatentCount(surveyConfig.szUnitCode, hallWebserviceConfig.handprintService.dataType, "", minusDateTime(kssj), jssj)
     info("latentCount number:{}",latentCount)
-    if(latentCount.toInt > 1){
+    if(latentCount > 1){
       var nIndex = surveyConfig.nSeq
       val step = 9
       Range(nIndex, latentCount,10).foreach{i =>
